+++ dmrtd-docs/docusaurus.config.ts
@@ -14,10 +14,14 @@ const config: Config = {
   },
 
   // Set the production url of your site here
-  url: "https://your-docusaurus-site.example.com",
+  url: "https://privacybydesign.github.io/",
   // Set the /<baseUrl>/ pathname under which your site is served
   // For GitHub pages deployment, it is often '/<projectName>/'
-  baseUrl: "/",
+  baseUrl: "/dmrtd/",
+  trailingSlash: false,
+  organizationName: "privacybydesign",
+  projectName: "dmrtd",
+  deploymentBranch: "gh-pages",
 
   onBrokenLinks: "throw",
   onBrokenMarkdownLinks: "warn",
