+++ dmrtd-docs/package.json
@@ -7,12 +7,13 @@
     "start": "docusaurus start",
     "build": "docusaurus build",
     "swizzle": "docusaurus swizzle",
-    "deploy": "docusaurus deploy",
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
-    "typecheck": "tsc"
+    "typecheck": "tsc",
+    "predeploy": "docusaurus build",
+    "deploy": "gh-pages -d build"
   },
   "dependencies": {
     "@docusaurus/core": "3.8.1",
